+++ package.json
@@ -2,7 +2,7 @@
   "name": "gherking",
   "version": "0.0.3",
   "description": "Simple pre-compiler for Gherkin feature files",
-  "main": "lib/index.js",
+  "main": "lib/index.ts",
   "bin": {
     "gherkin-precompiler": "bin/precompiler.js",
     "precompile": "bin/precompiler.js"
